+++ ruins/apps/weather.py
@@ -1,14 +1,17 @@
+from faulthandler import disable
 from typing import List, Callable, Tuple
 import streamlit as st
 import xarray as xr     # TODO: these references should be moved to DataManager
 import pandas as pd     # TODO: these references should be moved to DataManager
 import numpy as np
 import matplotlib.pyplot as plt
+from plotly.express.colors import named_colorscales
 
 from ruins.plotting import plt_map, kde, yrplot_hm
-from ruins.components import data_select, topic_select
+from ruins.components import data_select, model_scale_select, topic_select
 from ruins.core import build_config, debug_view, DataManager, Config
 from ruins.core.cache import partial_memoize
+from ruins.plotting.climate_parcoords import climate_projection_parcoords
 from ruins.processing.sdm import SDM
 
 
@@ -172,9 +175,117 @@ def _reduce_weather_data(dataManager: DataManager, name: str, variable: str, tim
         return df[station]       
 
 
+def climate_data_selector(dataManager: DataManager, config: Config, it: int = 0, variable: str = 'T', expander_container = st.sidebar, layout: str = 'columns', **kwargs):
+    """Handles the selection and display of one paralell coordinates plot"""
+    # create a unique key
+    key = f'rcp_reference_{it}'
+    title = f'Select Station #{it + 1}' if config.lang == 'en' else f'Station #{it + 1} auswählen'
+
+    # get the container
+    container = kwargs['container'] if 'container' in kwargs else st
+
+    # if this is not the first iteration, we pre-select an item for rcp
+    if it > 0:
+        st.session_state[key] = 'rcp45'
+
+    # make the data selection
+    data_select.rcp_selector(
+        dataManager,
+        config,
+        title=title,
+        expander_container=expander_container,
+        elements='__all__',
+        layout=layout,
+        RCP_KEY=key,
+        allow_skip=False
+    )
+
+    # get the reference from config
+    ref = config[key]
+
+    # make the data sub-selection
+    if ref == 'weather':
+        data = dataManager['weather'].read()
+        drngx = (1980, 2000)
+    else:
+        data = dataManager['cordex_coast'].read()
+        drngx = (2050, 2070)
+
+    # filter for rcps
+    if ref in ('rcp26', 'rcp45', 'rcp85'):
+        data = data.filter_by_attrs(RCP=ref)
+
+    # create the data range slider
+    drng = [pd.to_datetime(data.isel(time=0, vars=1).time.values).year, pd.to_datetime(data.isel(time=-1, vars=1).time.values).year]
+    datarng = expander_container.slider('Date Range', drng[0], drng[1], drngx, key=f'dr{it}')
+
+    # switch the variable
+    if variable == 'T':
+        afu = np.mean
+    elif variable == 'Tmin':
+        afu = np.min
+    elif variable == 'Tmax':
+        afu = np.max
+
+    # aggregate
+    dyp = data.sel(vars=variable).to_dataframe().resample('1M').apply(afu)
+    dyp = dyp.loc[(dyp.index.year>=datarng[0]) & (dyp.index.year<datarng[1]),dyp.columns[dyp.columns!='vars']]
+
+    # plot
+    fig = climate_projection_parcoords(data=dyp, colorscale=kwargs.get('colorscale', 'electric'))
+    container.plotly_chart(fig, use_container_width=True)
+
+# TODO refactor the plots into the plotting module
+def climate_plots(dataManager: DataManager, config: Config, expander_container = st.sidebar):
+    """
+    """
+    cliproj = config['climate_scale']
+
+
+    # TODO: build this into a Component ?
+    if cliproj=='Regional':
+        st.warning('Sorry, we currently have issues with the Regional model data. Please come back later.')
+        st.stop()
+        regaggs = ['North Sea Coast', 'Krummhörn',  'Niedersachsen', 'Inland']
+        regagg = st.sidebar.selectbox('Spatial aggregation:', regaggs)
+
+        if regagg=='North Sea Coast':
+            climate = xr.load_dataset('data/cordex_coast.nc')
+            climate.filter_by_attrs(RCP='rcp45')
+        elif regagg=='Krummhörn':
+            climate = xr.load_dataset('data/cordex_krummh.nc')
+
+    # TODO: Refactor this part - similar stuff is used in weather explorer
+    navi_vars = ['Maximum Air Temperature', 'Mean Air Temperature', 'Minimum Air Temperature']
+    navi_var = expander_container.radio("Select variable:", options=navi_vars)
+    if navi_var[:4] == 'Mini':
+        vari = 'Tmin'
+        ag = 'min'
+    elif navi_var[:4] == 'Maxi':
+        vari = 'Tmax'
+        ag = 'max'
+    else:
+        vari = 'T'
+        ag = 'mean'
+
+    # add plots as needed
+    num_of_plots = st.sidebar.number_input('# of datasets to compare', min_value=1, max_value=5, value=1)
+
+    for it in range(int(num_of_plots)):
+        # create expander
+        plot_expander = st.expander(f'Temperatur (°C) Monthly {ag}', expanded=(it==num_of_plots - 1))
+        left, right = plot_expander.columns((2, 8))
+        left.markdown('### Options')
+        opt = left.container()
+        # add the colorbar as option
+        cmap = left.selectbox(f'Plot #{it + 1} Colorbar', options=named_colorscales(), format_func=lambda l: l.capitalize())
+
+        # add the Parcoords plot
+        climate_data_selector(dataManager, config, it=it, variable=vari, colorscale=cmap, expander_container=opt, container=right)
+
+
 def warming_data_plotter(dataManager: DataManager, config: Config):
     weather: xr.Dataset = dataManager['weather'].read()
-    climate = dataManager['cordex_coast'].read()
     statios = list(weather.keys())
     stat1 = config['selected_station']
 
@@ -325,18 +436,97 @@ def warming_data_plotter(dataManager: DataManager, config: Config):
                 fig = yrplot_hm(sr=data2, ref=ref_yr2, ag=ag, lang=config.lang)
                 fig.update_layout(title=f"{stat2} {navi_var} anomaly to {ref_yr2[0]}-{ref_yr2[1]}")
                 plot_area.plotly_chart(fig, use_container_width=True)
-                
-                # old matplotlib include
-                #plt.title(stat2 + ' ' + navi_var + ' anomaly to ' + str(ref_yr2[0]) + '-' + str(ref_yr2[1]))
-                #st.pyplot(fig)
 
-        # Re-implement this as a application wide service
-        # expl_md = read_markdown_file('explainer/stripes_m.md')
-        # st.markdown(expl_md, unsafe_allow_html=True)
+
+def quick_access_buttons(config: Config, container = st.sidebar):
+    """Add quick access button to skip parts of the Weather explorer"""
+    # get the current stage
+    stage = config.get('quick_access')
+
+    l, r = container.columns(2)
+
+    # switch the cases
+    if stage == 'climate':
+        go_weather = l.button('Weather explorer')
+        go_climate = False
+        go_idx = r.button('Climate indices')
+    elif stage == 'index':
+        go_weather = l.button('Weather explorer')
+        go_climate = r.button('Climate explorer')
+        go_idx = False
+    else:
+        go_weather = False
+        go_climate = l.button('Climate explorer')
+        go_idx = r.button('Climate indices')
+
+    # check if the Weather explorer is needed
+    if go_weather:
+        st.session_state.quick_access = 'weather'
+        st.experimental_rerun()
+    
+    # check if the Climate explorer is needed
+    if go_climate:
+        if 'include_climate' in st.session_state:
+            del st.session_state['include_climate']
+        st.session_state.quick_access = 'climate'
+        st.experimental_rerun()
+    
+    # check if the Climate indices are needed
+    if go_idx:
+        if 'include_climate' in st.session_state:
+            del st.session_state['include_climate']
+        st.session_state.quick_access = 'index'
+        st.experimental_rerun()
+
+
+def weather_stage(dataManager: DataManager, config: Config, data_expander=st.sidebar):
+    # Story mode - go through each setting
+    # update session state with current data settings
+    data_expander = st.sidebar.expander('Data selection', expanded=True)
+    data_select.data_select(dataManager, config, expander_container=data_expander, container=st)
+    
+    
+    # build the app
+    st.header('Weather Data Explorer')
+    # TODO: move this text also into story mode?
+    st.markdown('''In this section we provide visualisations to explore changes in observed weather data. Based on different variables and climate indices it is possible to investigate how climate change manifests itself in different variables, at different stations and with different temporal aggregation.''',unsafe_allow_html=True)
+
+    warming_data_plotter(dataManager, config)
+
+
+def climate_stage(dataManager: DataManager, config: Config):
+    # Story mode - go through each setting
+
+    # update session state with current settings
+    # get model scale
+    option_container = st.sidebar.expander('OPTIONS', expanded=True)
+    model_scale_select.model_scale_selector(dataManager, config, expander_container=option_container)
+    
+    # run main visualization
+    climate_plots(dataManager, config, expander_container=option_container)
+
+
+def index_stage(dataManager: DataManager, config: Config, data_expander=st.sidebar):
+    # Story mode - go through each setting
+    # update session state with current data settings
+    data_expander = st.sidebar.expander('Data selection', expanded=True)
+    data_select.selected_station_selector(dataManager, config, expander_container=data_expander)
+    data_select.rcp_selector(dataManager, config, expander_container=data_expander)
+
+    # run actual visualization
+    climate_indices(dataManager, config)
 
 
 def main_app(**kwargs):
     """Describe the params in kwargs here
+
+    The main app has three 'stages': 
+      
+      * learning about weather
+      * learning about climate
+      * condensing info into climate indices
+
+    
     """
     # build the config and dataManager from kwargs
     url_params = st.experimental_get_query_params()
@@ -346,32 +536,28 @@ def main_app(**kwargs):
     st.set_page_config(page_title='Weather Explorer', layout=config.layout)
     debug_view.debug_view(dataManager, config, debug_name='DEBUG - initial state')
 
-    # Story mode - go through each setting
-    # update session state with current data settings
-    data_expander = st.sidebar.expander('Data selection', expanded=True)
-    data_select.data_select(dataManager, config, expander_container=data_expander, container=st)
-    
-    # topic selector
-    topic_expander = st.sidebar.expander('Topic selection', expanded=True) # here we would use control_policy ('show', 'hide) if we want to keep it
-    topic_select.topic_select(config=config, expander_container=topic_expander, container=st)
-    
-    # build the app
-    st.header('Weather Data Explorer')
-    # TODO: move this text also into story mode?
-    st.markdown('''In this section we provide visualisations to explore changes in observed weather data. Based on different variables and climate indices it is possible to investigate how climate change manifests itself in different variables, at different stations and with different temporal aggregation.''',unsafe_allow_html=True)
+    # check if a stage was set
+    if not config.has_key('quick_access'):
+        st.session_state.quick_access = 'weather'
+    stage = config['quick_access']
 
-    # topic is set, now run the correct plot
-    topic = config['current_topic']
+    # add the skip buttons
+    btn_expander = st.sidebar.expander('QUICK ACCESS', expanded=True)
+    quick_access_buttons(config, container=btn_expander)
 
-    if topic == 'Warming':
-        warming_data_plotter(dataManager, config)
-    
-    elif topic == 'Weather Indices':
-        climate_indices(dataManager, config)
 
+    # -------------
+    # Weather Stage
+    if stage == 'weather':
+        weather_stage(dataManager, config)
+    elif stage == 'climate':
+        climate_stage(dataManager, config)
+    elif stage == 'index':
+        index_stage(dataManager, config)
     else:
-        st.error(f'{topic} is not a valid topic here. Please tell the developer.')
-
+        st.error(f"We received weird data. A quick_access='{stage}' does not exist. Please contact the developer.")
+        st.stop()
+        
     # end state debug
     debug_view.debug_view(dataManager, config, debug_name='DEBUG - finished app')
